+++ app/models/organisation_settings.rb
@@ -22,10 +22,8 @@ class OrganisationSettings < Settings
   attribute :default_begins_at_time, :string, default: -> { '08:00' }
   attribute :default_ends_at_time, :string, default: -> { '16:00' }
   attribute :locales, array: true, default: -> { I18n.available_locales.map(&:to_s) }
-  # attribute :conflicting_occupancy_types, array: true, default: -> { Occupancy::OCCUPANCY_TYPES.keys - %i[free] }
-  attribute :occupied_occupancy_states, array: true, default: lambda {
-                                                                BookingStates.occupied_occupancy_able.keys.map(&:to_s)
-                                                              }
+  attribute :occupied_occupancy_states, array: true, default: -> { occupied_occupancy_able_booking_states }
+  attribute :predefined_salutation_form, :string
 
   validates :tentative_occupancy_color, :occupied_occupancy_color,
             :closed_occupancy_color, format: { with: Occupancy::COLOR_REGEX }, allow_blank: true
@@ -37,6 +35,8 @@ class OrganisationSettings < Settings
             :invoice_payment_deadline, :deposit_payment_deadline, :deadline_postponable_for, :upcoming_soon_window,
             :payment_overdue_deadline,
             numericality: { less_than_or_equal: 5.years, greater_than_or_equal: 0 }
+
+  validates :predefined_salutation_form, inclusion: { in: Tenant.salutation_forms.keys.map(&:to_s) }, allow_blank: true
 
   def occupancy_colors
     {
@@ -53,5 +53,9 @@ class OrganisationSettings < Settings
 
   def manage_transition_to_states(organisation)
     organisation.booking_flow_class.successors['initial'].map { BookingStates.all.fetch(_1.to_sym) }.compact_blank
+  end
+
+  def self.occupied_occupancy_able_booking_states
+    BookingStates.occupied_occupancy_able.keys.map(&:to_s)
   end
 end