+++ app/params/manage/tarif_params.rb
@@ -4,8 +4,8 @@ module Manage
   class TarifParams < ApplicationParams
     def self.permitted_keys
       %i[type label unit price_per_unit ordinal tarif_group accountancy_account pin
-         prefill_usage_method prefill_usage_booking_question_id
-         minimum_usage_per_night minimum_usage_total vat] +
+         prefill_usage_method prefill_usage_booking_question_id vat
+         minimum_usage_per_night minimum_usage_total minimum_price_per_night minimum_price_total] +
         I18n.available_locales.map { |locale| ["label_#{locale}", "unit_#{locale}"] }.flatten +
         [{ associated_types: [],
            selecting_conditions_attributes: BookingConditionParams.permitted_keys + %i[id _destroy],