+++ app/params/public/tenant_params.rb
@@ -4,7 +4,7 @@ module Public
   class TenantParams < ApplicationParams
     def self.permitted_keys
       %i[first_name last_name street_address zipcode city email birth_date country_code nickname phone
-         address_addon locale]
+         address_addon locale salutation_form]
     end
   end
 end