+++ app/serializers/manage/tenant_serializer.rb
@@ -2,7 +2,7 @@
 
 module Manage
   class TenantSerializer < ApplicationSerializer
-    fields :salutation_name, :first_name, :last_name, :street_address, :nickname, :name,
+    fields :salutation, :salutation_form, :first_name, :last_name, :street_address, :nickname, :name,
            :address_addon, :zipcode, :city, :email, :full_name, :contact_info, :phone,
            :address_lines, :full_address_lines, :birth_date, :country_code, :locale
   end